LayoutTests/platform/efl/accessibility/image-map1-expected.txt

 
This tests that you can reach the links within an image map.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.childAtIndex(0).role is 'AXRole: AXLink'
FAIL body.childAtIndex(0).title should be AXTitle: Link1. Was AXTitle: .
PASS body.childAtIndex(1).role is 'AXRole: AXLink'
FAIL body.childAtIndex(1).title should be AXTitle: Link2. Was AXTitle: .
PASS body.childAtIndex(2).role is 'AXRole: AXLink'
FAIL body.childAtIndex(2).title should be AXTitle: Link3. Was AXTitle: .
PASS successfullyParsed is true

TEST COMPLETE
